Move the remaining bits of mozzi_rand.cpp to other locations
While at it, use explict precision types.
