+++ MozziGuts_impl_RP2040.hpp
@@ -267,7 +267,7 @@ void stopMozzi() {
 ////// END audio output code //////
 
 //// BEGIN Random seeding ////////
-void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+void MozziRandPrivate::autoSeed() {
 #warning Automatic random seedings is not implemented on this platform
 }
 //// END Random seeding ////////
