+++ MozziGuts_impl_STM32.hpp
@@ -140,7 +140,7 @@ void stopMozzi() {
 //// END AUDIO OUTPUT code ///////
 
 //// BEGIN Random seeding ////////
-void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+void MozziRandPrivate::autoSeed() {
   // Unfortunately the internal temp sensor on STM32s does _not_ appear to create a lot of noise.
   // Ironically, the calls to calibrate help induce some random noise. You're still fairly likely to produce two equal
   // random seeds in two subsequent runs, however.
@@ -150,12 +150,12 @@ void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
     uint32_t ci;
   } conv;
   conv.cf = adc.readTemp();
-  *x=*x^conv.ci;
+  x=x^conv.ci;
   adc.calibrate();
   conv.cf = adc.readTemp();
-  *y=*y^conv.ci;
+  y=y^conv.ci;
   adc.calibrate();
   conv.cf = adc.readTemp();
-  *z=*z^conv.ci;
+  z=z^conv.ci;
 }
 //// END Random seeding ////////
