+++ MozziGuts_impl_TEENSY.hpp
@@ -114,7 +114,7 @@ void stopMozzi() {
 //// END AUDIO OUTPUT code ///////
 
 //// BEGIN Random seeding ////////
-void autoRandomSeeds(uint32_t *x, uint32_t *y, uint32_t *z) {
+void MozziRandPrivate::autoSeed() {
 #warning Automatic random seedings is not implemented on this platform
 }
 //// END Random seeding ////////
